requirements.txt
# ── Core framework ──────────────────────────────────────────────────────────
streamlit>=1.32.0
flask>=3.0.0

# ── LangChain ecosystem ─────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.9
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.2.0

# ── LangGraph ───────────────────────────────────────────────────────────────
langgraph>=0.2.0

# ── Vector store ────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Document loaders ────────────────────────────────────────────────────────
pypdf>=4.0.0

# ── Embeddings (Gemini Embedding 2) ─────────────────────────────────────────
google-generativeai>=0.8.0

# ── Data / validation ───────────────────────────────────────────────────────
pydantic>=2.0.0
python-dotenv>=1.0.0
